Struct unbound::Answer

pub struct Answer(_);

Wraps ub_result. The result of DNS resolution and validation of a query.

Methods

impl Answer

fn qname(&self) -> &str

Returns original question's name.

fn qtype(&self) -> u16

Returns original question's qtype.

fn qclass(&self) -> u16

Returns original question's qclass.

fn datas(&self) -> Datas

Returns an iterator over answer record datas.

fn canonname(&self) -> Option<&str>

Returns canonical name of result, if any.

fn rcode(&self) -> u16

Returns additional error code in case of no data.

fn answer(&self) -> Option<&[u8]>

Returns answer packet, if any.

fn havedata(&self) -> bool

Returns true if there is data.

fn nxdomain(&self) -> bool

Returns true if there is no data because a name does not exist.

fn secure(&self) -> bool

True if result is secure.

fn bogus(&self) -> bool

True if a security failure happened.

fn why_bogus(&self) -> Option<&str>

String error if response is bogus.

fn ttl(&self) -> u32

Number of seconds the result is valid.